Move notifications from ProgramFragment to MainActivity
